Added file controller

+++ WebApplication1/Controllers/ArtistController.cs
@@ -1,8 +1,6 @@
 ﻿using AutoMapper;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using System.Collections.Generic;
-using System.Threading.Tasks;
 using WebApplication1.Data.Entities;
 using WebApplication1.Interfaces;
 using WebApplication1.Models.DTOs.Artist;
@@ -14,11 +12,13 @@ public class ArtistController : ControllerBase
 {
     private readonly IArtistService _artistService;
     private readonly IMapper _mapper;
+    private readonly IFileService _fileService;
 
-    public ArtistController(IArtistService artistService, IMapper mapper)
+    public ArtistController(IArtistService artistService, IMapper mapper, IFileService fileService)
     {
         _artistService = artistService;
         _mapper = mapper;
+        _fileService = fileService;
     }
 
     [HttpGet]
@@ -26,8 +26,8 @@ public class ArtistController : ControllerBase
     public async Task<ActionResult<IEnumerable<ArtistInformationDto>>> GetArtists()
     {
         var artists = await _artistService.GetAllArtists();
-        var artistDtosResponse = _mapper.Map<IEnumerable<ArtistInformationDtoResponse>>(artists);
-        return Ok(artistDtosResponse);
+        var artistDtoResponse = _mapper.Map<IEnumerable<ArtistInformationDtoResponse>>(artists);
+        return Ok(artistDtoResponse);
     }
 
     [HttpGet("{id}")]
@@ -35,9 +35,14 @@ public class ArtistController : ControllerBase
     public async Task<ActionResult<ArtistInformationDto>> GetArtistById(int id)
     {
         var artist = await _artistService.GetArtistById(id);
-        var artistDtosResponse = _mapper.Map<IEnumerable<ArtistInformationDtoResponse>>(artist);
 
-        return Ok(artistDtosResponse);
+        if (artist == null)
+        {
+            return NotFound();
+        }
+
+        var artistDtoResponse = _mapper.Map<ArtistInformationDtoResponse>(artist);
+        return Ok(artistDtoResponse);
     }
 
     [HttpPost]
@@ -46,53 +51,71 @@ public class ArtistController : ControllerBase
     {
         if (artistDto.Image == null)
         {
-            return BadRequest("Image file is requred");
+            return BadRequest("Image file is required");
         }
 
-        string imagePath = SaveFile(artistDto.Image, "images", "artists");
+        string imagePath = _fileService.SaveFile(artistDto.Image, "images");
 
         var artist = _mapper.Map<ArtistInformation>(artistDto);
         artist.Image = imagePath;
+        artist.DateRegistered = DateTime.Now;
 
-        
-        await _artistService.AddArtist(artist); 
+        await _artistService.AddArtist(artist);
 
         return Ok();
     }
 
     [HttpPut("{id}")]
     [Authorize(Roles = "Admin")]
-    public async Task<IActionResult> UpdateArtist(int id, [FromForm] ArtistInformation updatedArtist)
+    public async Task<ActionResult<ArtistInformationDtoResponse>> UpdateArtist(int id, [FromForm] ArtistInformationDto updatedArtistDto)
     {
-        try
+        var existingArtist = await _artistService.GetArtistById(id);
+
+        if (existingArtist == null)
         {
-            await _artistService.UpdateArtist(id, updatedArtist);
-            return Ok();
+            return NotFound();
         }
-        catch (ArgumentException ex)
+
+        string imagePath = null;
+
+        if (updatedArtistDto.Image != null)
         {
-            return NotFound(ex.Message);
+            _fileService.DeleteFile(existingArtist.Image);
+
+            imagePath = _fileService.SaveFile(updatedArtistDto.Image, "images");
+            existingArtist.Image = imagePath;
         }
+        else
+        {
+            imagePath = existingArtist.Image;
+        }
+
+        updatedArtistDto.DateRegistered = existingArtist.DateRegistered;
+
+        _mapper.Map(updatedArtistDto, existingArtist);
+
+        if (!string.IsNullOrEmpty(imagePath))
+        {
+            existingArtist.Image = imagePath;
+        }
+
+        await _artistService.UpdateArtist(id, existingArtist);
+
+        var updatedArtistDtoResponse = _mapper.Map<ArtistInformationDtoResponse>(existingArtist);
+
+        return Ok(updatedArtistDtoResponse);
     }
 
     [HttpDelete("{id}")]
-    public async Task<IActionResult> DeleteArtist(int id)
+    public async Task<ActionResult> DeleteArtist(int id)
     {
+        var artist = await _artistService.GetArtistById(id);
+        if (artist == null)
+        {
+            return NotFound();
+        }
+        _fileService.DeleteFile(artist.Image);
         await _artistService.DeleteArtist(id);
         return Ok();
-    }
-
-    private string SaveFile(IFormFile file, string subfolder1, string subfolder2)
-    {
-        var folderPath = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot", subfolder1, subfolder2);
-        var fileName = Guid.NewGuid().ToString() + Path.GetExtension(file.FileName);
-        var filePath = Path.Combine(folderPath, fileName);
-
-        using (var stream = new FileStream(filePath, FileMode.Create))
-        {
-            file.CopyTo(stream);
-        }
-
-        return $"/{subfolder1}/{subfolder2}/{fileName}";
     }
 }
